+++ src/appInstantiator.js
