+++ src/fingerprinter.js
