+++ src/visitor.js
